+++ packages/docs/src/stories/Button.stories.tsx
@@ -5,6 +5,10 @@ import { ArrowRight } from 'phosphor-react'
 export default {
   title: 'Form/Button',
   component: Button,
+  tags: ['autodocs'],
+  parameters: {
+    layout: 'centered',
+  },
   args: {
     children: 'Submit',
     color: 'primary',