+++ packages/docs/src/stories/Checkbox.stories.tsx
@@ -4,6 +4,10 @@ import type { Meta, StoryObj } from '@storybook/react'
 export default {
   title: 'Form/Checkbox',
   component: Checkbox,
+  tags: ['autodocs'],
+  parameters: {
+    layout: 'centered',
+  },
   args: {
     name: 'name',
   },