+++ packages/docs/src/stories/TextArea.stories.tsx
@@ -4,6 +4,10 @@ import type { Meta, StoryObj } from '@storybook/react'
 export default {
   title: 'Form/TextArea',
   component: TextArea,
+  tags: ['autodocs'],
+  parameters: {
+    layout: 'centered',
+  },
   args: {
     name: 'name',
     placeholder: 'Escreva um comentário',
@@ -24,7 +28,26 @@ export default {
         type: 'boolean',
       },
     },
+    error: {
+      control: {
+        type: 'text',
+      },
+    },
   },
 } as Meta<TextAreaProps>
 
 export const Default: StoryObj<TextAreaProps> = {}
+
+export const Error: StoryObj<TextAreaProps> = {
+  args: {
+    label: 'Message',
+    error: 'Field is required'
+  },
+}
+
+export const Required: StoryObj<TextAreaProps> = {
+  args: {
+    label: 'Message',
+    required: true
+  },
+}