+++ packages/docs/src/stories/TextField.stories.tsx
@@ -5,6 +5,10 @@ import { MagnifyingGlass } from 'phosphor-react'
 export default {
   title: 'Form/TextField',
   component: TextField,
+  tags: ['autodocs'],
+  parameters: {
+    layout: 'centered',
+  },
   args: {
     name: 'name',
   },
@@ -17,6 +21,11 @@ export default {
     disabled: {
       control: {
         type: 'boolean',
+      },
+    },
+    error: {
+      control: {
+        type: 'text',
       },
     },
   },
@@ -34,5 +43,19 @@ export const HasIcon: StoryObj<TextFieldProps> = {
   args: {
     icon: <MagnifyingGlass />,
     placeholder: 'Search',
+  },
+}
+
+export const Error: StoryObj<TextFieldProps> = {
+  args: {
+    label: 'First Name',
+    error: 'Field is required'
+  },
+}
+
+export const Required: StoryObj<TextFieldProps> = {
+  args: {
+    label: 'Last Name',
+    required: true
   },
 }