+++ packages/react/src/components/TextArea/index.tsx
@@ -1,17 +1,21 @@
 import { ComponentProps, ElementRef, forwardRef } from 'react'
-import { Container, Input, Text } from './styles'
+import { Container, HelperText, Input, Text } from './styles'
 
-export interface TextAreaProps extends ComponentProps<typeof Input> {
+export interface TextAreaProps extends Omit<ComponentProps<typeof Input>, 'error'> {
   label?: string
+  required?: boolean
+  error?: string
 }
 
 export const TextArea = forwardRef<ElementRef<typeof Input>, TextAreaProps>(
-  ({ label, ...props }: TextAreaProps, ref) => {
+  ({ label, error, required, ...props }: TextAreaProps, ref) => {
     return (
       <Container>
-        {label && <Text htmlFor={props.name}>{label}</Text>}
+        {label && <Text htmlFor={props.name} error={!!error}>{label} {required && <span>*</span>}</Text>}
 
-        <Input ref={ref} {...props} id={props.name} />
+        <Input ref={ref} {...props} id={props.name} error={!!error} />
+
+        {error && <HelperText>{error}</HelperText>}
       </Container>
     )
   },