+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @acampe-ui/react
 
+## 2.4.1
+
+### Patch Changes
+
+- added error props in TextField component
+  added error props in TextArea component
+  password field is now optional in TextField component
+
 ## 2.4.0
 
 ### Minor Changes